Progress on training.

+++ src/export.cpp
@@ -22,10 +22,9 @@ FileWriter::FileWriter(int w, int h, NeuralNet nn, string filename) :
 		file << w << ' ' << h << endl;
 	}
 	//Generate data
-	for (int j = 0; j < h; ++j) {
-		for (int i = 0; i < w; ++i) {
-			Matrix out = nn.eval(Matrix(1, 2, new long double[2] { i,
-					j }));
+	for (long double j = 0; j < h; ++j) {
+		for (long double i = 0; i < w; ++i) {
+			Matrix out = nn.eval(Matrix(1, 2, new long double[2] { i, j }));
 			file << i << ' ' << j << ' ' << out.get(0, 0) << ' '
 					<< out.get(0, 1) << ' ' << out.get(0, 2) << endl;
 		}
